ASoC: jz4740-i2s: Add codec runtime switching support
The I2S controller on many Ingenic SoCs has a register bit to
select between an internal codec integrated with the SoC or an
external discrete codec. Only one codec can be used at a time,
but some boards have both codecs wired up and can switch between
them at runtime.

Because the playback and capture routes are tied together the
switch cannot be represented as a normal DAPM widget. Instead
the switch is implemented as a normal control that dynamically
adds & removes DAPM routes.

The device tree is used to define which codecs are present
with ingenic,internal-codec and ingenic,external-codec flags.
When there is only one codec present, the selection will be
programmed in the register and cannot be changed at runtime.
No controls or DAPM widgets will be added to the component,
so userspace won't be able to select an invalid configuration.

When both codecs are present, the codec switch control will be
added along with the DAPM widgets and routes needed to handle
switching.

Since the JZ47xx SoCs don't have a card driver, the connection
between the codec streams and AIF widgets for the codec ports
needs to be defined using DAPM routing in the device tree when
using two codecs; eg. for simple-audio-card:

    sound {
        compatible = "simple-audio-card";
        simple-audio-card,routing =
            "IntCodec Playback", "I2S_INT_TX",
            "IntCodec Capture", "I2S_INT_RX",
            "ExtCodec Playback", "I2S_EXT_TX",
            "ExtCodec Capture", "I2S_EXT_RX";

        simple-audio-card,dai-link@0 {
            format = "i2s";

            cpu {
                sound-dai = <&aic>;
            };

            codec {
                  prefix = "IntCodec";
                  sound-dai = <&codec>;
            };
        };

        simple-audio-card,dai-link@1 {
            format = "i2s";

            cpu {
                sound-dai = <&aic>;
            };

            codec {
                  prefix = "ExtCodec";
                  sound-dai = <&external_codec>;
            };
        };
    };

Boards that only use one codec don't need to describe routing
explicitly, since they don't have the AIF widgets and can only
be routed in one way.