dt-bindings: ingenic,aic: Add bindings for codec paths
The AIC on Ingenic SoCs can be connected to an internal codec
integrated with the SoC or an external codec. The active codec
can be selected at runtime with an AIC register. Playback and
capture is only possible for the active codec. In other words
only one codec can be used at a time, but it's possible to switch
the active codec at runtime.

The DT needs to describe which codec paths are actually usable
on a given platform. This allows paths for unusable codecs to
be pruned away so they can't be selected by userspace.