dt-bindings: rtc: ingenic: Add rtc-valid-value
The RTC on Ingenic SoCs detects power loss events by comparing
a persistent register with a constant. Because the constant is
arbitrary, it should be specified in the device tree so OSs can
agree on the value (for example, Ingenic's BSP kernels use a
different constant than the upstream Linux driver).