rtc: ingenic: Parse rtc-valid-value from the device tree
The RTC valid value is a constant used to detect if a power loss
has occurred. Historically Linux used 0x12345678, but that's just
an arbitrary number. The device tree should specify this value to
be properly OS-independent, so take the value from the DT if it's
available, otherwise fall back on the old value.