scsi: ses: Avoid reporting errors for simple subenclosures
Currently the SES driver reports bogus errors when plugging
in a drive that exposes only a single, simple subenclosure:

[  432.713731] scsi 3:0:0:0: Direct-Access     WD       Elements 2667    2007 PQ: 0 ANSI: 6
[  432.714127] scsi 3:0:0:1: Enclosure         WD       SES Device       2007 PQ: 0 ANSI: 6
...
[  432.716508] scsi 3:0:0:1: Attached scsi generic sg2 type 13
...
[  439.897020] scsi 3:0:0:1: Wrong diagnostic page; asked for 1 got 8
[  439.897023] scsi 3:0:0:1: Failed to get diagnostic page 0x1
[  439.897025] scsi 3:0:0:1: Failed to bind enclosure -19

According to the SES specification, simple subenclosures always
return diagnostic page 8 no matter what page was requested, so
the device is permitted to page 8 here and nothing is wrong.

To avoid polluting the kernel logs with bogus errors, the first
diagnostic page read bypasses the usual page code check. If it
returns page 8 the device is assumed to be a simple subenclosure
and no enclosure device is created. Simple subenclosures only
return a vendor specific status byte in page 8 and don't support
any other pages, so they can't support the enclosure interface.