regulator: axp20x: Refactor conditionally supported regulators
Add the .is_regulator_supported() function to determine if a
particular regulator should be registered or not at runtime.

This gets rid of special cases for poly-phase regulators and
the AXP813's FLDO3, hiding the details behind a more general
interface and keeping the generic part of the driver simple.